requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
idna==3.10
psycopg2==2.9.9
pycparser==2.22
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
telebot==0.0.5
urllib3==2.2.3
